Move all the process checker classes in a dedicated package.
