+++ process-manager-api/src/main/java/org/orbisgis/processmanagerapi/check/ICheckDataBuilder.java
@@ -34,7 +34,7 @@
  * or contact directly:
  * info_at_ orbisgis.org
  */
-package org.orbisgis.processmanagerapi;
+package org.orbisgis.processmanagerapi.check;
 
 /**
  * Interface for the definition of which data are provided for the process check execution.
