+++ process-manager-api/src/main/java/org/orbisgis/processmanagerapi/check/IProcessCheck.java
@@ -34,9 +34,10 @@
  * or contact directly:
  * info_at_ orbisgis.org
  */
-package org.orbisgis.processmanagerapi;
+package org.orbisgis.processmanagerapi.check;
 
 import groovy.lang.Closure;
+import org.orbisgis.processmanagerapi.IProcess;
 
 import java.util.LinkedHashMap;
